float correction

+++ SRC/API-DATA-RETRIEVE.py
@@ -40,7 +40,7 @@ def create_tables(cursor):
     query = '''CREATE TABLE IF NOT EXISTS actors (
                 id INT PRIMARY KEY,
                 actor_name VARCHAR(100) NOT NULL)
-                popularity FLOAT(5, 3)
+                popularity DECIMAL (5,3)
                 '''
     cursor.execute(query)
